LPS-94642 Move to a more generic util since it's not a dto anymore
